:ghost: SoftError deprecated; addon error handling improved.
