+++ addon/adapter.go
@@ -42,7 +42,6 @@ type Field = binding.Field
 
 //
 // Error
-type SoftError = binding.SoftError
 type ResetError = binding.RestError
 type Conflict = binding.Conflict
 type NotFound = binding.NotFound
@@ -105,11 +104,8 @@ func (h *Adapter) Run(addon func() error) {
 			}
 		}
 		if err != nil {
-			if _, soft := err.(interface{ Soft() *SoftError }); !soft {
-				Log.Error(err, "Addon failed.")
-				os.Exit(1)
-			}
 			h.Failed(err.Error())
+			os.Exit(1)
 		}
 	}()
 	//
