+++ binding/error.go
@@ -8,28 +8,9 @@ import (
 )
 
 //
-// SoftError A "soft" anticipated error.
-type SoftError struct {
-	Reason string
-}
-
-func (e *SoftError) Error() (s string) {
-	return e.Reason
-}
-
-func (e *SoftError) Is(err error) (matched bool) {
-	_, matched = err.(*SoftError)
-	return
-}
-
-func (e *SoftError) Soft() *SoftError {
-	return e
-}
-
-//
 // RestError reports REST errors.
 type RestError struct {
-	SoftError
+	Reason string
 	Method string
 	Path   string
 	Status int
